+++ packages/dnd/jest.config.mjs
@@ -4,6 +4,7 @@ import common from '../../jest.common.mjs';
 export default {
   ...common,
   displayName: 'dnd',
+  coveragePathIgnorePatterns: [...common.coveragePathIgnorePatterns, 'dropEffectListener'],
   transformIgnorePatterns: ['/node[_]modules/(?!react-dnd|dnd-core|@react-dnd)'],
   setupFilesAfterEnv: [...common.setupFilesAfterEnv, './jestSetup.ts'],
 };